requirements.txt
RPi.GPIO>=0.6.3
numpy>=1.14.1
opencv-python
